python/requirements.txt
flask==3.0.0
flask-cors==4.0.0
kafka-python==2.0.2
grpcio==1.59.0
grpcio-tools==1.59.0
